Add tests for crash timing predictability (#117)
We are planning to rework how collisions work (#93). Instead of saving all the pixels that the Kurves have drawn and checking for overlap (with a special case for 45-degree draws), we are planning to save the positions of the middle of each 3-by-3 square (“head”) we draw, and use those positions for collision detection.

When working on that, we realized that a naive implementation resulted in Kurves dying too early when approaching a wall or other Kurve at an angle. This made the gameplay unpredictable to experienced players, and deviated from the original game. The reason this happened was because we were using the internal floating point positions of Kurves for collision detection, resulting in a mismatch between what the player(s) would see and what would actually be happening behind the scenes.

This PR adds regression tests for this kind of crash timing predictability, to ensure we don’t regress in this area when reworking the collisions.

Co-authored-by: Simon Alling <alling.simon@gmail.com>