Add test case for cutting corners (#118)
In the original game, it's possible – and quite common – to "paint over" already painted pixels without crashing. As far as we know, it can only ever happen in a so-called 45-degree draw, which is an explicit special case in our game; see `hitbox` in the `World` module.

In our planned revamped collision detection (#93), we're not planning to have such a special case; it should just be a consequence of the algorithm.

This PR adds a basic regression test for cutting a corner, along with a sanity-check test verifying that crashing into a corner head-on (i.e. not cutting it) _is_ lethal.

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>