Add test case for the perfect overpainting (#126)
I've never actually seen this in the original game; it's just something I assume is possible based on my understanding of its collision handling. Whether that's correct or not, it's how our clone works, so why not have a regression test? Then it will also be very explicit if we deliberately change the collision handling so that it's not possible anymore.