Use `requestAnimationFrame`-based main loop (#150)
Today, we simulate real time with `Time.every` (so presumably `setInterval` under the hood). Its [docs] are crystal clear though:

> **This function is not for animation.** Use the [`onAnimationFrame`][af] function for that sort of thing! It syncs up with repaints and will end up being much smoother for any moving visuals.
>
> [af]: https://package.elm-lang.org/packages/elm/browser/latest/Browser-Events#onAnimationFrame

So using `onAnimationFrame` (based on `requestAnimationFrame`) is exactly what we do in this PR, except we use `onAnimationFrameDelta` for convenience. The implementation is based on @IceCreamYou's [`MainLoop.js`][MainLoop], which was used in the legacy JavaScript version of this game, and whose inner workings are described in detail in [this excellent blog post].

And yes: the game does indeed run much more smoothly, at least on a 60 Hz monitor. However, the smoothness seems _worse_ at 120 Hz, at least when not running the browser in fullscreen. This seems to be the case in both Chrome and Firefox on my main PC with Windows, but it's very inconsistent and hard to reproduce reliably. I have also observed severe stuttering on my laptop with a 60 Hz screen and Ubuntu 20.04. I really don't know …

For reference: the `AroundTheWorld` test case is great for testing animation smoothness.

## About the diff

The "starting point" of the changeset is that the `Time.every` subscription is replaced by a `Browser.Events.onAnimationFrameDelta` one. The corresponding message variant is renamed from `GameTick` to `AnimationFrame` and augmented to hold the frame delta of the current frame and any leftover time from the previous frame.

The next step is that the application of `Game.reactToTick` is moved from `update` into the new `MainLoop.consumeAnimationFrame` function.

All other changes are made in response to the changes above. In particular:

  * `ActiveGameState` is modified so that the `Moving` variant can hold any leftover frame time that needs to be accounted for in the next animation frame.
  * `TickResult` is modified so that the `RoundKeepsGoing` variant can hold any type of information, because then the `RoundKeepsGoing`/`RoundEnds` structure can be used both when we have a `LeftoverFrameTime` and a `Tick`, and when we don't.
  * `tickResultToGameState` is modified not to take a `Tick` anymore, instead taking a `TickResult ( LeftoverFrameTime, Tick, MidRoundState )`, because such a value is readily available at the call site.

[docs]: https://github.com/elm/time/blob/dc3b75b7366e59b99962706f7bf064d3634a4bba/src/Time.elm#L414C1-L416C44
[MainLoop]: https://github.com/IceCreamYou/MainLoop.js/tree/247e7c41fe4bfa7e15ff4cc524d56056feffd306
[this excellent blog post]: http://www.isaacsukin.com/news/2015/01/detailed-explanation-javascript-game-loops-and-timing

💡 `git show --color-words=' -> |Game\.reactToTick|Time\.every .+always <|\w+|.`

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>