Move static files to `_site` directory (#160)
The legacy JavaScript version that was superseded in `42a9090c20b6c6b1ac7c5adbf7494e4d6ffb682d` was deployed to GitHub Pages by committing and pushing compiled code to `master` using the `deploy` npm script removed in #158. The last version that was deployed seems to have been `28a4e5142ebe7c0329669166252ef9605242bad8`, in April 2019 …

It's about time we start deploying the Elm version, and I think a GitHub Actions workflow is a better solution – see #157. I think that will be easier if the built site ends up in a single directory, instead of in the repo root. I've decided to go with `_site/` because [it's the default value for the `path` input of `actions/upload-pages-artifact`][docs] (as of v4.0.0).

What this PR does:

  * Moves all static assets (`index.html` and `resources/`) to `_site/` – no build step needed.
  * Makes `elm-watch` output to `_site/`.
  * Makes `sass` output to `_site/css/`.
  * Makes `esbuild` serve `_site/`.
  * Adapts the reference to `ZATACKA.js` in `index.html`.

[docs]: https://github.com/actions/upload-pages-artifact/tree/v4.0.0?tab=readme-ov-file#inputs-

💡 `git show --color-words='build|.'`