Add script for manipulating original game (#164)
This landmark PR adds a basic proof-of-concept script that modifies the memory of the original game to set up a custom scenario. It includes a small DSL for expressing scenarios, and the minimum scaffolding for staging a scenario in an already running `ZATACKA.EXE` inside DOSBox.

An example scenario is included which can be staged like this:

  1. Install [scanmem](https://github.com/scanmem/scanmem) version 0.17.

  2. Temporarily [disable address space layout randomization (ASLR)][aslr]:

     ```bash
     echo 0 | sudo tee /proc/sys/kernel/randomize_va_space
     ```

     Otherwise, the game's state won't end up at a predictable location in memory.

  3. Set `memsize=1` in DOSBox's config file (`~/.dosbox/dosbox-0.74-3.conf` or similar). With the default value of 16, the game's state still won't end up at a predictable location in memory.

  4. Launch the original game as usual.

  5. Join with Red, Yellow and Green, and start the game.

  6. When the Kurves have started moving:

     ```bash
     sudo ./tools/scenario.py `pgrep dosbox`
     ```

Red, Yellow and Green should immediately teleport and start moving perfectly horizontally.

It works because the game stores the players' x coordinates, y coordinates and directions as 32-bit floats starting at `0x7FFFD8010FF6` (at least on my ThinkPad X1 Carbon). This command prints those three arrays (indexed by player ID, with Red = 0):

    sudo scanmem `pgrep dosbox` --errexit --command "option endianness 1;dump 0x7fffd8010ff6 72;exit"

Each player's direction seems to be represented in counter-clockwise radians, with 0 somewhat counterintuitively for me being down, π / 2 being right, etc. It doesn't seem to be normalized to any specific interval; I've seen it grow to above 20 and shrink to below –10.

This is a major breakthrough in our efforts to reverse-engineer the original game. Based on [the groundwork] laid by @lydell and me in #93, @Swij and I came up with this script during a trip to [Vroeger Was Alles Beter: Classics Reloaded][vwab]. Turns out that the front seats of a Tesla Model 3, the Rødby–Puttgarden ferry and a Dutch hotel room are quite decent environments for reverse engineering an old DOS game.

[aslr]: https://askubuntu.com/questions/318315/how-can-i-temporarily-disable-aslr-address-space-layout-randomization/318476#318476
[the groundwork]: https://github.com/SimonAlling/kurve/issues/93#issuecomment-3417461145
[vwab]: https://vwab.nl/