Run `sudo` early in scenario script (#167)
I constantly make the mistake of running the scenario script without having done `sudo true` or similar first. What happens then is that the scenario isn't staged at all, and when I go back to the terminal, it's stuck on `[sudo] password for alling:`.

This PR improves the situation by running `sudo` early, so that the password prompt appears before DOSBox is launched in front of the terminal.