Simplify and standardize test cases (#174)
Today, all our test cases follow the same structure: initialize a round using `roundWith`, pass the resulting `RoundInitialState` to `expectRoundOutcome`, and tell it what expectations to place on the round when it is over. It's quite verbose, repetitive and fragile.

This PR standardizes the test cases, removes boilerplate and simplifies them: instead of placing arbitrary expectations on various aspects of the finished round (which tend to always be very similar anyway), we describe the expected ending in a standardized, unambiguous way where it's not possible to forget anything.

One concrete benefit of the new way is that a `RoundOutcome` doesn't contain any `Expectation` anymore, which makes it possible to move the expected outcomes into the corresponding `TestScenarios` modules, should we want to. They can also be serialized, should the need arise.

This improvement was triggered by the recent breakthroughs in #93. I think they will allow us, with some degree of automation, to define test cases based on the actual behavior of the original game. Then I think it will be extra important for test cases to be expressed as simply as possible.

## About `AliveKurve`

One might wonder why `AliveKurve` doesn't have a `theDrawingPositionItNeverMadeItTo` field like `DeadKurve`. Well, we don't have any test case that checks where the winner was when the round ended, so it's completely unnecessary. Admittedly, we don't have any test case that checks who won either, so the `id` field is also unnecessary by the same logic. But I think it makes test cases more readable if it's explicitly stated who should win.

## About `aliveAtTheEnd`

One might wonder why the `aliveAtTheEnd` field has type `List AliveKurve` and not something like `Maybe AliveKurve`. The latter would perhaps result in nicer-looking test cases, but it doesn't match our model: the type checker doesn't know that there cannot be more than one Kurve alive at the end of a round. (Note that there are _zero_ Kurves alive at the end of a single-player round.)