Limit DOSBox memory size in scenario script (#178)
As mentioned in #164, #165 and #170, the scenario script relies on `memsize=1` being set in the user's DOSBox config file (`~/.dosbox/dosbox-0.74-3.conf` or similar). This PR removes that assumption, instead forcing `memsize=1` automatically.

## About the diff

`-userconf` is important because without it, the user's config file is completely ignored. We probably don't want it to be ignored, because the user may have useful settings there, such as window resolution, scaler settings, etc. In any case, ignoring it isn't in scope for this PR.

💡 `git show --color-words=.`