Check if ASLR is disabled in scenario script (#179)
As mentioned in #164 and #170, the scenario script doesn't work if address space layout randomization is enabled, because then the game's state doesn't end up at a predictable location in memory.

Note that ASLR is a security feature that shouldn't be permanently disabled. It should only be _temporarily_ disabled while working with the scenario script and then enabled again.

How to disable ASLR:

    echo 0 | sudo tee /proc/sys/kernel/randomize_va_space

How to enable ASLR:

    echo 2 | sudo tee /proc/sys/kernel/randomize_va_space