Fix and enforce formatting in tools/ (#185)
VS Code auto-formats Python code with Ruff on my laptop, but I haven't set that up on my main PC, so I accidentally added non-formatted code in #183, and VS Code keeps formatting it on my laptop.

This PR fixes the formatting and adds a CI check that runs [Ruff] on `tools/`.

[Ruff]: https://docs.astral.sh/ruff/