Make scenario script work in WSL on my Windows PC (#187)
As mentioned in #180 and #183, I haven't been able to make the scenario script work in WSL (Ubuntu 24.04.2) on my Windows PC with `memsize=1`, and it doesn't work on my Ubuntu laptop (Ubuntu 24.04.3) _without_ `memsize=1`.

This PR effectively forces a choice of `memsize` to be made when running the scenario script. It does so by making the scenario script take the `memsize`-overriding DOSBox config file as a command-line argument, adding an ostensibly WSL-specific config file, and renaming the existing file to make it more obvious which one to pick.

I want to emphasize that I really don't know what this discrepancy between my computers stems from – maybe it doesn't have anything to do with WSL at all – so do note that the names of the DOSBox config files are purely based on a wild guess. In any case, it would have felt even stranger to name them `dosbox-on-my-ubuntu-laptop.conf` and `dosbox-in-wsl-on-my-windows-pc.conf` or similar …

## How to use the scenario script as of this PR

### On my laptop

    ./tools/scenario.py docs/original-game/ZATACKA.EXE 7fffd8010ff6 tools/dosbox-linux.conf

### In WSL on my main PC

    ./tools/scenario.py docs/original-game/ZATACKA.EXE 7fffc1c65ff6 ./tools/dosbox-wsl.conf