Align direction representation with original game (#192)
As documented in #191, we've recently learned that our internal representation of a Kurve's direction is different from that in the original game. We have directions zeroed around rightwards/east, but the original game has them zeroed around downwards/south.

In order to make it easier to align the observable mechanics of our clone with those of the original game (for example by staging scenarios using `tools/scenario.py` and then creating test cases based on their respective outcomes), I think we should try to align our internal representation with the original game as much as possible.

💡 `git show --color-words='-- .+|cos|sin|def set_direction|conventional|.'`