Scenariofy "Crash timing" test cases (#194)
Continuing where #181 left off, this PR aligns the test cases in `crashingIntoKurveTimingTests` with the pattern of describing the scenario in a `TestScenarios` module. One concrete benefit of that is that it becomes easier to visualize the scenario, as described in #144.

Unlike all other existing test scenarios, this one is parameterized over Red's y coordinate, so its `spawnedKurves` has type `Float -> List Kurve` instead of `List Kurve`.

💡 `git show --color-moved --color-moved-ws=allow-indentation-change`