Make `find_and_focus_dosbox` exit on failure (#206)
This PR makes `find_and_focus_dosbox` print an error message and exit itself, instead of returning `None`, if it fails to find the DOSBox window.

Note that since we use `--sync` (#204), "fails to find the DOSBox window" refers to problems like xdotool not being installed or the arguments passed to it being invalid. If the xdotool command is correct but the DOSBox window isn't found, then `find_and_focus_dosbox` will just hang indefinitely.

💡 `git show --color-words='return None|.'`