Check if DOSBox is already open in scenario script (#208)
One of the most recurring mistakes we find ourselves making is running the scenario script without having closing DOSBox after the previous run. Then xdotool fails to find and focus the DOSBox window, and the scenario isn't staged.

This PR adds a fail-early check to minimize confusion and provide faster feedback. It does so by extracting the finding logic into its own `find_dosbox` function, which is then reused as the foundation of the check.

The finding logic is slightly different between the two use cases:

  * `--sync` means "wait until the window is found", which makes absolutely no sense when we expect _not_ to find any window, i.e. in `check_that_dosbox_is_not_already_open`.
  * `--onlyvisible` reads to me like open DOSBox windows may not be found if they are not "visible" (whatever that means), but I think we do want to find them in `check_that_dosbox_is_not_already_open`.

💡 `git show --ignore-space-change --color-moved --color-moved-ws=allow-indentation-change`