Fix flaky DOSBox focusing in scenario script (#209)
## Problem

Very intermittently, but often enough to be annoying (on my Ubuntu laptop), I've been getting these errors when running the scenario script:

  * ```
    ❌ Couldn't find the DOSBox window.
    ```

  * ```
    X Error of failed request:  BadWindow (invalid Window parameter)
      Major opcode of failed request:  20 (X_GetProperty)
      Resource id in failed request:  0x1a00017
      Serial number of failed request:  26
      Current serial number in output stream:  26
    ```

I printed `res.stdout` and `res.stderr` in `find_dosbox` and saw that the first error above was accompanied by this:

    X Error of failed request:  BadWindow (invalid Window parameter)
      Major opcode of failed request:  3 (X_GetWindowAttributes)
      Resource id in failed request:  0x811788
      Serial number of failed request:  386
      Current serial number in output stream:  387

The second error stems, I think, from the `windowactivate` and/or `windowfocus` commands.

## Solution

I asked ChatGPT 5 and got this explanation:

> This usually means: _by the time_ `xdotool` tried to talk to that window, it was already gone.

As a "common pattern that triggers this", ChatGPT mentioned this:

> You grab the id of a popup/dialog/splash that disappears quickly.

So I assume a temporary DOSBox window with a matching name is opened and then immediately replaced with the "real" one. But sometimes we happen to get the ID of the first window, which is gone by the time we try to do something with it.

This PR makes the script wait until DOSBox should have completed its launch before trying to find and focus it.