Enforce automation in scenario script (#211)
This PR makes the scenario script's automated mode (added in #165) the _only_ way to run it. We see no use for the non-automated mode anymore and therefore don't want to maintain it.

That is, as of this PR, the only way to run the scenario script is by specifying the path to the original game. It's not possible to specify the process ID of an already running DOSBox anymore.

💡 `git show --ignore-all-space --color-words='process_id|.'`