Rewrite original-game scenario DSL in Elm (#223)
The time has come! Once we started experimenting with gdb instead of scanmem (#199), I started yearning for a language more suitable for DSL implementation than Python. Luckily enough, we have one close at hand … ^^

This PR replaces the scenario DSL in `scenario.py` with a functionally equivalent – but architecturally quite different – Elm implementation.

A brief summary of the most interesting modules:

  * `MemoryLayout` contains our understanding of the memory layout of the original game (see #93).
  * `ModMem` describes how to modify memory, without involving any implementation details.
  * `Scanmem` compiles a `ModMem` representation to a string that can be passed to scanmem's `--command` flag.
  * `ScenarioCLI` handles I/O.
  * `ScenarioCore` describes a scenario and computes a `ModMem` representation from it.
  * `TheScenario` contains the same information as the removed `SCENARIO` constant.

Future improvements hopefully unlocked by this:

  * Replacing scanmem with gdb.
  * Unifying test-case scenarios and original-game scenarios.

## Notable implementation aspects

### `0x` prefix

Support for a `0x` prefix in the base address is implicit today, because Python's `hex` function supports it. In this PR, it's instead explicit, because `Integer.fromHexString` doesn't support it.

### Participating players

Today, there's no need to worry about invalid player IDs like `42`, because that would require obviously incorrect code like this:

```python
PlayerId(42): {
    "x": 100,
    "y": 100,
    "direction": pi,
},
```

From the perspective of `scenario.py`, getting the player IDs from Elm as a list of integers makes it slightly more likely for an invalid ID to sneak in.

But that's not a problem in practice. Both today and as of this PR, an invalid ID will cause `scenario.py` to fail like this:

    ValueError: 42 is not a valid PlayerId

### Python–Elm integration

The scenario compilation result is passed from Elm to Python as JSON. If the result isn't valid JSON, then `scenario.py` will fail gracefully. However, if the result is valid JSON but doesn't parse to a `CompilationResultAsJson`, then all bets are off.

Therefore, the Python–Elm integration is deliberately designed to make it easy to understand, using tools like `git grep`, how it all fits together. In particular, field names are unique, so that e.g. `git grep participatingPlayersById` reveals how the `participatingPlayersById` field is produced in Elm and consumed in Python, and nothing else.

### Blue

Just like today, if Blue participates in the scenario, the scenario script will fail like this in the lobby:

    ❌ Blue (the player) isn't supported yet.

The reason it's not supported is that we haven't yet figured out how to reliably send mouse clicks to the DOSBox window.

### CI

In CI, the new Elm program is compiled both directly in the Dockerfile and indirectly by the scenario script (#227). The new steps in the Validation workflow are based on their counterparts in the Deployment workflow.

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>