Remove accidentally added ScenarioAPI module (#229)
Ouch. Despite my best attempts to keep #223 clean, I managed to include a leftover artifact from the development phase: the file `ScenarioAPI.elm`.

For anyone wondering why `elm-review` didn't report the module as unused: it has an export named `main`, so it's considered an entry point and therefore not an unused module.