Replace scanmem with gdb in scenario script (#231)
As mentioned in #223, it's time to replace scanmem with gdb.

## Problem

While scanmem was a great first step to get the scenario script up and running in #164, it's not really powerful enough for our use case, because it cannot be used to modify the memory with precise timing. This has two disadvantages:

  * The scenario script has to wait until the Kurves have started moving before staging the scenario, so the staged scenario only represents the Kurves _coming from_ the specified positions, not them _spawning_ there (see #199). This makes it impossible to stage scenarios where the spawn positions are significant (see #197), and it introduces a confusing discrepancy between test-case scenarios and staged scenarios even when the exact spawn positions aren't relevant to the outcome.

  * Sometimes the scenario script happens to modify the memory in the middle of a tick, causing the Kurves to go out of sync with each other. For example, the scenario in `TheScenario.elm` normally ends with Red as the loser, Yellow as the runner-up and Green as the winner, because that's the order in which they are checked for collision with the wall (see #196). But it _sometimes_ ends differently, because we _sometimes_ happen to modify the memory when one or two of the players have been updated, but not all.

## Solution

This PR improves the situation. With gdb's `watch` command, we can modify each address immediately after the game itself has written to it, eliminating the timing problem and _almost_ enabling us to stage the scenario before the Kurves have started moving.

However, we've still only managed to move the Kurves just _after_ they have spawned. So their original spawn positions will be occupied and drawn, and #199 is still present: they only _come from_ the positions specified in the scenario. So we're not _quite_ able to spawn Kurves for real yet, but this is still a great improvement.

### Logging

I've completely suppressed all output from gdb in the scenario script; see comment in `scenario.py`. Check `gdb-log.txt` to see what gdb did.

### Missing pixel in original head

For some reason, one pixel is missing in the upper left corner of each head at the original spawn locations – except for Red's. For example:

    ⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛🟨🟨⬛⬛
    ⬛⬛🟨🟨🟨⬛⬛
    ⬛⬛🟨🟨🟨⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛

It doesn't seem like a visual artifact; that pixel really seems to be unoccupied, as demonstrated by this scenario (which does play out identically on both of my computers, which both have DOSBox 0.74-3):

```elm
[ ( Red
  , { x = 112
    , y = 136
    , direction = 0
    }
  )
, ( Yellow
  , { x = 500
    , y = 300
    , direction = 0
    }
  )
]
```

Red crashes like this:

    ⬛⬛🟥🟥🟥⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥🟨🟨⬛⬛
    ⬛⬛⬛⬛🟨🟨🟨⬛⬛
    ⬛⬛⬛⬛🟨🟨🟨⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛⬛⬛

### `compileCore`

Unlike `write` in scanmem, gdb's `watch` has a nested structure to it: what to do after modifying the memory is described as a continuation in the middle of the `watch` block, somewhat like when using `setTimeout` in JavaScript.

Therefore, it's not possible – or at least not easy – to implement `compileCore` using `List.map` – we need `List.foldr`. My understanding is that the reason it's `foldr` and not `foldl` is that to compile `ModMemCmd` number 𝑛, one must first have compiled number 𝑛 + 1, because it's the continuation part of number 𝑛.

### `exit` after last watchpoint

If we don't tell gdb to exit after hitting the last (innermost) watchpoint, then it remains attached to DOSBox, rendering `show-game-state` unusable, because it cannot attach to DOSBox while gdb is attached.

### `delete $bpnum`

If we don't delete the watchpoints after hitting them, gdb will fail like this:

    Could not insert hardware watchpoint 5.
    Could not insert hardware breakpoints:
    You may have requested too many hardware breakpoints/watchpoints.

My understanding is that [x86 supports 4 hardware watchpoints] using registers DR0–DR3. We always need at least 6 watchpoints (at least 2 players × 3 state components per player).

### `show-game-state.sh`

Without the changes in `show-game-state.sh`, it prints this error while gdb is attached to DOSBox (i.e. from the splash screen until the scenario has been staged):

    ⚠️  Read memory failed. Maybe the game is currently starting.

That's the default warning we print if we haven't identified a specific reason why scanmem cannot read the memory.

[x86 supports 4 hardware watchpoints]: https://en.wikipedia.org/wiki/X86_debug_register#DR0_to_DR3

💡 `git show -M15 --color-words='"option endianness .+"|"write .+|"exit"|"--[a-z]+"|[A-Z_]+|[a-z]+|.'`

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>