Include comments in compiled gdb program (#234)
The gdb programs we produce are quite difficult to read. That was true with scanmem as well (i.e. before #231), but scanmem programs cannot contain comments or even newlines as far as I know, so we couldn't do much about it.

This PR improves the situation by adding comments describing the purpose of each section.

💡 `git show --color-words=.`