Replace comments with printing in gdb program (#235)
The comments added in #234 are great, but they only make the compiled program itself more readable – the log it produces is still really hard to follow. This PR improves the situation by replacing the comments with `print` commands and also adding matching `print` commands at the end of every logical section.

💡 `git show --color-words='\w+|.'`