Check scenario before compiling it (#236)
## Problem

This scenario had me scratching my head for an embarrassing number of minutes:

```elm
    [ ( Red
      , { x = 200
        , y = 50
        , direction = pi / 2
        }
      )
    , ( Yellow
      , { x = 200
        , y = 100
        , direction = pi / 2
        }
      )
    , ( Green
      , { x = 200
        , y = 150
        , direction = pi / 2
        }
      )
    , ( Orange
      , { x = 200
        , y = 200
        , direction = pi / 2
        }
      )
    , ( Pink
      , { x = 200
        , y = 250
        , direction = pi / 2
        }
      )
    ]
```

Red, Yellow and Green work fine. Orange is moved correctly but doesn't have its direction set. Pink isn't manipulated at all.

This is the last thing we print in `gdb-log.txt`:

    $27 = "⏳ 🟧 Set Orange's direction"
    Hardware watchpoint 14: *(float*)0x7fffc1c6602e

The reason the gdb program gets stuck is that the players are specified in the wrong order: Orange should come before Green.

The reason the gdb program gets stuck specifically on setting Orange's direction is that, unlike Orange's x and y, its direction never changes again after we missed the chance to overwrite it at spawn time.

## Solution

This PR adds some basic checks for mistakes that are easy to make in a scenario.

There is of course a more [Making Impossible States Impossible™]-ish way to prevent both duplicate and incorrectly ordered players:

```elm
type alias Scenario =
    { red : Maybe PlayerState
    , yellow : Maybe PlayerState
    , orange : Maybe PlayerState
    , green : Maybe PlayerState
    , pink : Maybe PlayerState
    , blue : Maybe PlayerState
    }
```

Maybe we'll go there in the future, but this is good enough for now.

### Why not just sort the scenario?

We could. But there is value in standardizing the way in which scenarios are expressed, and making the relation between scenario and compiled gdb program more obvious.

[Making Impossible States Impossible™]: https://www.youtube.com/watch?v=IcgmSRJHu_8

💡 `git show --color-words='scenario|.'`